Introduce ARDUINO_INLINE_API macro
The ARDUINO_INLINE_API macro is used when defining
Arduino APIs that are compiled as 'static inline'
functions for the real hardware.  However, when
compiling for the emulator, we define the macro as
empty, making the symbol to appear in the emulator
shared library.  However, note that in most cases
they will be mangled C++ names, and must be called
as such from the Python side.